pyproject.toml
[tool.poetry]
name = "history_book"
version = "0.1.0"
description = "A sophisticated Python application for ingesting, storing, and querying historical books using Weaviate vector database"
authors = ["Christopher Burke <imcjb8@gmail.com>"]
readme = "README.md"
packages = [{ include = "history_book", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.14.1 || >3.14.1,<4.0"
pymupdf = ">=1.25.5,<2.0.0"
unstructured = ">=0.17.2,<0.18.0"
tiktoken = "^0.9.0"
weaviate-client = "^4.15.0"
pandas = "^2.3.0"
matplotlib = "^3.10.3"
marimo = {extras = ["recommended"], version = "^0.14.12"}
langchain = "^1.0"
langchain-openai = "^1.0"
langchain-anthropic = "^1.0"
langgraph = "^1.0"
langgraph-checkpoint = "^2.0"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic-settings = "^2.12.0"
networkx = "^3.6.1"
plotly = "^6.5.0"
pyvis = "^0.3.2"
scikit-learn = "^1.8.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.1"
notebook = "^7.5.1"
ipykernel = "^7.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
extend-exclude = [
    "*.ipynb",
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "notebooks/*.py" # marimo notebooks
]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.11
target-version = "py311"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F", "I", "UP", "B", "SIM", "PL", "TID252"]
ignore = ["PLR2004", "PLR1714", "PLR0912", "SIM108", "B008", "PLR0915", "PLR0913"]  # magic values, multiple comparisons, too many branches, ternary, function call in arg, too many statements, too many args

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
docstring-code-format = false